Feature add authorization jwt

+++ SocialMedia/Controllers/PostController.cs
@@ -1,13 +1,17 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using SocialMedia.models.DTO;
-using SocialMedia.models.DTO.PostReaction;
-using SocialMedia.models.DTO.Posts;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using SocialMedia.DTO;
+using SocialMedia.DTO.PostReaction;
+using SocialMedia.DTO.Posts;
+using SocialMedia.models;
 using SocialMedia.Services.Interfaces;
+using System.Security.Claims;
 
 namespace SocialMedia.Controllers
 {
     [ApiController]
     [Route("[controller]")]
+    [Authorize]
     public class PostController : ControllerBase
     {
         private readonly IPostsServices _postService;
@@ -19,30 +23,59 @@ namespace SocialMedia.Controllers
            
         }
 
+       
+
         [HttpPost]
         [Route("GetPostsList")]
-        public async Task<ActionResult<PostResponseDTO>> Get([FromBody]PostsFilterDTO filters,Guid? UserId)
+        public async Task<ActionResult<PostResponseDTO>> Get([FromBody] PostsFilterDTO filters)
         {
-            var postList = await _postService.GetAllPostsAsync(filters, UserId);
+
+
+            string LoggedInUser = await _postService.VerifyUser(User);
+
+            if (LoggedInUser == string.Empty) {
+                return Unauthorized();
+            }
+
+
+            var postList = await _postService.GetAllPostsAsync(filters, Guid.Parse(LoggedInUser));
             return Ok(postList);
         }
 
         [HttpPost("/CreatePost")]
-        public async Task<IActionResult> CreatePost([FromBody] CreatePostDTO postDto)
+        public async Task<IActionResult> CreatePost([FromBody] CreatePostDTO newPost)
         {
-            var createdPost = await _postService.CreatePostAsync(postDto);
+            string LoggedInUser = await _postService.VerifyUser(User);
+
+            if (LoggedInUser == string.Empty)
+            {
+                return Unauthorized();
+            }
+
+            Guid UserId = Guid.Parse(LoggedInUser);
+
+            PostResponseDTO createdPost = await _postService.CreatePostAsync(newPost,UserId);
             return CreatedAtAction(
-                nameof(GetPostByID),
-                new { PostId = createdPost.Id },
-                createdPost
-            );
+                    nameof(GetPostByID),
+                    createdPost
+                );
         }
 
         [HttpPost]
         [Route("React")]
         public async Task<IActionResult> ReactToPost([FromBody] ReactToPostDTO Reaction)
         {
-           await _postService.PostReaction(Reaction);
+
+            string LoggedInUser = await _postService.VerifyUser(User);
+
+            if (LoggedInUser == string.Empty)
+            {
+                return Unauthorized();
+            }
+
+            var UserId = Guid.Parse(LoggedInUser);
+
+            await _postService.PostReaction(Reaction,UserId);
             
             return NoContent();
         }
@@ -50,6 +83,14 @@ namespace SocialMedia.Controllers
         [HttpGet("{Id:guid}")]
         public async Task<IActionResult> GetPostByID([FromRoute] Guid Id)
         {
+            string LoggedInUser = await _postService.VerifyUser(User);
+
+            if (LoggedInUser == string.Empty)
+            {
+                return Unauthorized();
+            }
+
+
             var post = await _postService.GetPostByIdAsync(Id);
             if (post == null)
             {
@@ -61,6 +102,29 @@ namespace SocialMedia.Controllers
         [HttpDelete("{Id:guid}")]
         public async Task<IActionResult> DeletePost([FromRoute] Guid Id)
         {
+
+            string LoggedInUser = await _postService.VerifyUser(User);
+
+            if (LoggedInUser == string.Empty)
+            {
+                return Unauthorized();
+            }
+
+         
+
+            PostResponseDTO? post = await _postService.GetPostByIdAsync(Id);
+
+            if (post == null)
+            {
+                return NotFound();
+            }
+
+            if(post.UserId.ToString() != LoggedInUser) 
+            {
+                return Forbid();
+            }
+
+         
             await _postService.DeletePostAsync(Id);
             return NoContent();
         }
@@ -69,11 +133,29 @@ namespace SocialMedia.Controllers
         [HttpPut("{Id:guid}")]
         public async Task<IActionResult> UpdatePost([FromRoute] Guid Id, [FromBody] CreatePostDTO UpdatedPost)
         {
-            var post = await _postService.GetPostByIdAsync(Id);
+
+            string LoggedInUser = await _postService.VerifyUser(User);
+
+            if (LoggedInUser == string.Empty)
+            {
+                return Unauthorized();
+            }
+
+         
+
+            PostResponseDTO? post = await _postService.GetPostByIdAsync(Id);
+
             if (post == null)
             {
                 return NotFound();
             }
+
+            if (post.UserId.ToString() != LoggedInUser)
+            {
+                return Forbid();
+            }
+
+
             await _postService.UpdatePostAsync(Id, UpdatedPost);
             return NoContent();
         }
